Separate initS4sdkPipeline from runAsStage
Move initS4sdkPipeline to separate step so it is not run as a stage,
thus running in the default pod.
